requirements.txt
# PoliTrade 3.0 — Requirements

# ── Async HTTP (replaces requests in v2) ───────────────────────────────────────
httpx>=0.27.0

# ── HTML parsing ──────────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ── Schwab API ────────────────────────────────────────────────────────────────
schwab-py>=1.5.0

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0

# ── Market data (price fallback + historical) ─────────────────────────────────
yfinance>=0.2.40

# ── Utilities ─────────────────────────────────────────────────────────────────
pytz>=2023.3
python-dateutil>=2.8.2
rich>=13.7.0
